[WIP & CHORE] WebRTC 기능 구현중
WebRTC 관련 코드 작성,
방 생성 관련 코드 수정

+++ YogaYo/domain/src/main/java/com/d104/domain/model/ScoreUpdateMessage.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 @SerialName("score_update") // JSON 직렬화 시 type 필드 값으로 사용될 수 있음
 data class ScoreUpdateMessage(
-    val score: Int,
+    val score: Float,
     override val type: String = "score_update"
 ) : DataChannelMessage()
